pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = ""
authors = ["DarthZiel <Studerzhan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
djangorestframework = "^3.14.0"
channels = "^4.0.0"
daphne = "^4.0.0"
djangorestframework-simplejwt = "^5.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
